Quick Sort (Sorting Algorithm )
Step 1. Take a element and send to its Right position(same position after sorting)
Step 2. Check Condition that is make sure the left part of that array element is less than the element which I set to its right position(in step 1) and same right part element id greater than that element.
( <a | a | >a )